+++ src/ui/dialogs/generalsettingspanel.h
@@ -44,6 +44,7 @@ signals:
 private slots:
   void onBrowseHomeViewIcon();
   void onBrowseStartupVideo();
+  void onBrowseRetroarchConfig();
 
 private:
   void writeBack();